requirements.txt
alembic==1.7.7
apispec==5.1.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.0
Flask==2.1.0
flask-apispec==0.11.1
Flask-Bcrypt==1.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.1.1
six==1.16.0
SQLAlchemy==1.4.34
tomli==2.0.1
Unidecode==1.1.2
webargs==8.1.0
Werkzeug==2.1.0
zipp==3.7.0
